+++ packages/pigeon/platform_tests/alternate_language_test_plugin/ios/Classes/AlternateLanguageTestPlugin.m
@@ -258,20 +258,20 @@
 }
 
 - (void)callFlutterNoopWithCompletion:(void (^)(FlutterError *_Nullable))completion {
-  [self.flutterAPI noopWithCompletion:^(NSError *error) {
+  [self.flutterAPI noopWithCompletion:^(FlutterError *error) {
     completion(error);
   }];
 }
 
 - (void)callFlutterThrowErrorWithCompletion:(void (^)(id _Nullable,
                                                       FlutterError *_Nullable))completion {
-  [self.flutterAPI throwErrorWithCompletion:^(id value, NSError *error) {
+  [self.flutterAPI throwErrorWithCompletion:^(id value, FlutterError *error) {
     completion(value, error);
   }];
 }
 
 - (void)callFlutterThrowErrorFromVoidWithCompletion:(void (^)(FlutterError *_Nullable))completion {
-  [self.flutterAPI throwErrorFromVoidWithCompletion:^(NSError *error) {
+  [self.flutterAPI throwErrorFromVoidWithCompletion:^(FlutterError *error) {
     completion(error);
   }];
 }
@@ -279,7 +279,7 @@
 - (void)callFlutterEchoAllTypes:(AllTypes *)everything
                      completion:(void (^)(AllTypes *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoAllTypes:everything
-                     completion:^(AllTypes *value, NSError *error) {
+                     completion:^(AllTypes *value, FlutterError *error) {
                        completion(value, error);
                      }];
 }
@@ -292,7 +292,7 @@
   [self.flutterAPI sendMultipleNullableTypesABool:aNullableBool
                                             anInt:aNullableInt
                                           aString:aNullableString
-                                       completion:^(AllNullableTypes *value, NSError *error) {
+                                       completion:^(AllNullableTypes *value, FlutterError *error) {
                                          completion(value, error);
                                        }];
 }
@@ -300,7 +300,7 @@
 - (void)callFlutterEchoBool:(NSNumber *)aBool
                  completion:(void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoBool:aBool
-                 completion:^(NSNumber *value, NSError *error) {
+                 completion:^(NSNumber *value, FlutterError *error) {
                    completion(value, error);
                  }];
 }
@@ -308,7 +308,7 @@
 - (void)callFlutterEchoInt:(NSNumber *)anInt
                 completion:(void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoInt:anInt
-                completion:^(NSNumber *value, NSError *error) {
+                completion:^(NSNumber *value, FlutterError *error) {
                   completion(value, error);
                 }];
 }
@@ -316,7 +316,7 @@
 - (void)callFlutterEchoDouble:(NSNumber *)aDouble
                    completion:(void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoDouble:aDouble
-                   completion:^(NSNumber *value, NSError *error) {
+                   completion:^(NSNumber *value, FlutterError *error) {
                      completion(value, error);
                    }];
 }
@@ -324,7 +324,7 @@
 - (void)callFlutterEchoString:(NSString *)aString
                    completion:(void (^)(NSString *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoString:aString
-                   completion:^(NSString *value, NSError *error) {
+                   completion:^(NSString *value, FlutterError *error) {
                      completion(value, error);
                    }];
 }
@@ -333,7 +333,7 @@
                       completion:(void (^)(FlutterStandardTypedData *_Nullable,
                                            FlutterError *_Nullable))completion {
   [self.flutterAPI echoUint8List:aList
-                      completion:^(FlutterStandardTypedData *value, NSError *error) {
+                      completion:^(FlutterStandardTypedData *value, FlutterError *error) {
                         completion(value, error);
                       }];
 }
@@ -341,7 +341,7 @@
 - (void)callFlutterEchoList:(NSArray<id> *)aList
                  completion:(void (^)(NSArray<id> *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoList:aList
-                 completion:^(NSArray<id> *value, NSError *error) {
+                 completion:^(NSArray<id> *value, FlutterError *error) {
                    completion(value, error);
                  }];
 }
@@ -350,7 +350,7 @@
                 completion:(void (^)(NSDictionary<NSString *, id> *_Nullable,
                                      FlutterError *_Nullable))completion {
   [self.flutterAPI echoMap:aMap
-                completion:^(NSDictionary<NSString *, id> *value, NSError *error) {
+                completion:^(NSDictionary<NSString *, id> *value, FlutterError *error) {
                   completion(value, error);
                 }];
 }
@@ -359,7 +359,7 @@
                          completion:
                              (void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableBool:aBool
-                         completion:^(NSNumber *value, NSError *error) {
+                         completion:^(NSNumber *value, FlutterError *error) {
                            completion(value, error);
                          }];
 }
@@ -368,7 +368,7 @@
                         completion:
                             (void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableInt:anInt
-                        completion:^(NSNumber *value, NSError *error) {
+                        completion:^(NSNumber *value, FlutterError *error) {
                           completion(value, error);
                         }];
 }
@@ -377,7 +377,7 @@
                            completion:
                                (void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableDouble:aDouble
-                           completion:^(NSNumber *value, NSError *error) {
+                           completion:^(NSNumber *value, FlutterError *error) {
                              completion(value, error);
                            }];
 }
@@ -386,7 +386,7 @@
                            completion:
                                (void (^)(NSString *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableString:aString
-                           completion:^(NSString *value, NSError *error) {
+                           completion:^(NSString *value, FlutterError *error) {
                              completion(value, error);
                            }];
 }
@@ -395,7 +395,7 @@
                               completion:(void (^)(FlutterStandardTypedData *_Nullable,
                                                    FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableUint8List:aList
-                              completion:^(FlutterStandardTypedData *value, NSError *error) {
+                              completion:^(FlutterStandardTypedData *value, FlutterError *error) {
                                 completion(value, error);
                               }];
 }
@@ -404,7 +404,7 @@
                          completion:
                              (void (^)(NSArray<id> *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableList:aList
-                         completion:^(NSArray<id> *value, NSError *error) {
+                         completion:^(NSArray<id> *value, FlutterError *error) {
                            completion(value, error);
                          }];
 }
@@ -413,7 +413,7 @@
                         completion:(void (^)(NSDictionary<NSString *, id> *_Nullable,
                                              FlutterError *_Nullable))completion {
   [self.flutterAPI echoNullableMap:aMap
-                        completion:^(NSDictionary<NSString *, id> *value, NSError *error) {
+                        completion:^(NSDictionary<NSString *, id> *value, FlutterError *error) {
                           completion(value, error);
                         }];
 }