+++ packages/pigeon/platform_tests/shared_test_plugin_code/lib/integration_tests.dart
@@ -988,8 +988,6 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       expect(api.callFlutterNoop(), completes);
     });
 
-    // TODO(tarrinneal): add flutter api error handling and enable these tests,
-    // See: https://github.com/flutter/flutter/issues/118243
     testWidgets('errors are returned from non void methods correctly',
         (WidgetTester _) async {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
@@ -997,7 +995,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       expect(() async {
         await api.callFlutterThrowError();
       }, throwsA(isA<PlatformException>()));
-    }, skip: true);
+    },
+        // TODO(tarrinneal): Once flutter api error handling is added, enable these tests.
+        // See: https://github.com/flutter/flutter/issues/118243
+        skip: true);
 
     testWidgets('errors are returned from void methods correctly',
         (WidgetTester _) async {
@@ -1006,7 +1007,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       expect(() async {
         await api.callFlutterThrowErrorFromVoid();
       }, throwsA(isA<PlatformException>()));
-    }, skip: true);
+    },
+        // TODO(tarrinneal): Once flutter api error handling is added, enable these tests.
+        // See: https://github.com/flutter/flutter/issues/118243
+        skip: true);
 
     testWidgets('all datatypes serialize and deserialize correctly',
         (WidgetTester _) async {
@@ -1304,8 +1308,6 @@ class _FlutterApiTestImplementation implements FlutterIntegrationCoreApi {
   @override
   void noop() {}
 
-  // TODO(tarrinneal): update these to send error to platform,
-  // See: https://github.com/flutter/flutter/issues/118243
   @override
   Object? throwError() {
     throw FlutterError('this is an error');